requirements.txt
python-dotenv>=1.0.1
pandas>=2.2.2
pytest>=8.2.0
pytest-cov>=5.0.0
colorama>=0.4.6